+++ internal/pkg/http/api/constants.gen.go
@@ -10,5 +10,5 @@ const (
 	MaxMetadataEntries     int    = 16
 	MaxMetadataValueLength int    = 256
 	MaxNoteLength          int    = 2000
-	SpecVersion            string = "0.3.0"
+	SpecVersion            string = "0.4.0"
 )